+++ src/main/java/com/delivery/restaurant/cart/CartItemRepository.java
@@ -2,6 +2,6 @@ package com.delivery.restaurant.cart;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 
 }
